requirement fix

+++ 16-binary_tree_is_perfect.c
@@ -20,7 +20,7 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
 	height = binary_tree_height(tree);
 	size = binary_tree_size(tree);
 
-	power = _pow_recursion(2, height + 1);
+	power = (size_t)_pow_recursion(2, height + 1);
 	return (power - 1 == size);
 }
